% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_est_analytical.R
\name{get_est_analytical}
\alias{get_est_analytical}
\title{Generate point estimates with analytical confidence intervals from a model object}
\usage{
get_est_analytical(
  model,
  data = model$model,
  formula = model$formula,
  variable,
  alpha = 0.05,
  twotailed = TRUE
)
}
\arguments{
\item{model}{A model object}

\item{data}{A data frame (by default, \code{model$model} is used)}

\item{formula}{The model formula (by default, \code{model$formula} is used)}

\item{variable}{Column name for the group indicator}

\item{alpha}{Confidence level}

\item{twotailed}{Logical flag indicating whether to perform a two-tailed
hypothesis test}
}
\value{
Returns, as a named list, group-specific point estimates and all
pairwise differences with analytical confidence intervals.
}
\description{
Generate point estimates with analytical confidence intervals from a model object
}
