% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_est_analytical_continuous.R
\name{get_est_analytical_continuous}
\alias{get_est_analytical_continuous}
\title{Generate continuous estimates with analytical confidence intervals from a model object}
\usage{
get_est_analytical_continuous(
  model,
  data = model$model,
  formula = model$formula,
  group_variable,
  continuous_variable,
  seq_length = 21L,
  alpha = 0.05,
  twotailed = TRUE
)
}
\arguments{
\item{model}{A \code{lm} or \code{glm} model object}

\item{data}{A data frame (by default, \code{model$model} is used)}

\item{formula}{The model formula (by default, \code{model$formula} is used)}

\item{group_variable}{Column name for the group indicator}

\item{continuous_variable}{Column name for the continuous variable along which
group-specific conditional expectations will be computed}

\item{seq_length}{Length of the value sequence of \code{continuous_variable}
for the calculation of the continuous estimates}

\item{alpha}{Confidence level}

\item{twotailed}{Two-tailed (the default) or one-tailed hypothesis test}
}
\value{
Returns group-specific continuous estimates and all pairwise
differences with analytical 95% confidence intervals.
}
\description{
Generate continuous estimates with analytical confidence intervals from a model object
}
