% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_est_numerical.R
\name{get_est_numerical}
\alias{get_est_numerical}
\title{Generate point estimates with simulation-based, bootstrapped, or
Bayesian uncertainty intervals}
\usage{
get_est_numerical(
  model = NULL,
  data = model$model,
  formula = model$formula,
  variable,
  twotailed = FALSE,
  alpha = 0.95,
  n_draws = 1000L,
  seed = 20220315L,
  link = "identity",
  type = c("simulation", "bootstrap", "bayesian"),
  posterior_draws = NULL
)
}
\arguments{
\item{model}{A model object}

\item{data}{A data frame (by default, \code{model$model} is used)}

\item{formula}{The model formula (by default, \code{model$formula} is used)}

\item{variable}{Column name for the group indicator}

\item{twotailed}{Logical flag indicating whether to perform a two-tailed
hypothesis test}

\item{alpha}{Confidence level}

\item{n_draws}{Number of simulation draws or bootstrap resamples}

\item{seed}{Seed for reproducibility}

\item{link}{A character that specifies the name of the link function
or a user-supplied link function}

\item{type}{A character that specifies the type of the interval:
\code{"simulation"}, \code{"bootstrap"}, \code{"bayesian"}. Simulation-based
and bootstrapped confidence intervals are calculated as part of the function;
Bayesian credible intervals require the provision of \code{posterior_draws}
along with \code{data}, \code{formula}, and \variable{variable}.}

\item{posterior_draws}{A matrix of posterior draws (in rows) of the model
coefficients (columns). Only required and evaluated if \code{type == "bayesian"}.}
}
\value{
Returns, as a named list, group-specific point estimates and all
pairwise differences with numerical uncertainty intervals as well as the underlying
draws of these quantities of interest.
}
\description{
Generate point estimates with simulation-based, bootstrapped, or
Bayesian uncertainty intervals
}
