% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_est_numerical_flex.R
\name{get_est_numerical_flex}
\alias{get_est_numerical_flex}
\title{Generate point estimates with simulation-based, bootstrapped, or
Bayesian uncertainty intervals}
\usage{
get_est_numerical_flex(
  draws,
  variable,
  twotailed = FALSE,
  alpha = 0.05,
  type = c("simulation", "bootstrap", "bayesian")
)
}
\arguments{
\item{draws}{A matrix of numerical draws (in rows) of the quantities of interest
(columns).}

\item{variable}{Column name for the group indicator}

\item{twotailed}{Logical flag indicating whether to perform a two-tailed
hypothesis test}

\item{alpha}{Confidence level}

\item{type}{A character that specifies the type of the interval:
\code{"simulation"}, \code{"bootstrap"}, \code{"bayesian"}.}
}
\value{
Returns, as a named list, group-specific point estimates and all
pairwise differences with numerical uncertainty intervals as well as the underlying
draws of these quantities of interest.
}
\description{
Generate point estimates with simulation-based, bootstrapped, or
Bayesian uncertainty intervals
}
