% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_continuous_estimates.R
\name{plot_continuous_estimates}
\alias{plot_continuous_estimates}
\title{Generate (interactive) plots for multi-group comparisons of continuous estimates}
\usage{
plot_continuous_estimates(
  estimates,
  type = c("analytical", "simulation", "bootstrap", "bayesian"),
  variable,
  variable_label = NA,
  round_dec = 2,
  color_palette = "viridis",
  one_tailed_test = FALSE,
  p_val_threshold = 0.05,
  p_val_type = "p-value",
  interactive = FALSE
)
}
\arguments{
\item{estimates}{A \code{triviz_estimates} object generated by the package's suite
of \code{get_est} commands or a user-supplied object of equivalent structure (for
details, please see the vignette)}

\item{type}{Specifies how uncertainty estimates are derived (evaluated only when
\code{estimates} is a user-supplied object). Input must be one of
\code{"analytical"}, \code{"simulation"}, \code{"bootstrap"}, or \code{"bayesian"}.}

\item{variable}{Column name of the continuous predictor along which estimates will
be plotted}

\item{variable_label}{Label for the continuous predictor along which estimates will
be plotted}

\item{round_dec}{Number of decimal points for rounding of the x-axis}

\item{color_palette}{Character specifying the name of the color palette to be used.
Defaults to \code{"viridis"}.}

\item{one_tailed_test}{= (evaluated only when \code{estimates} is a user-supplied object)}

\item{p_val_threshold}{= (evaluated only when \code{estimates} is a user-supplied object)}

\item{p_val_type}{(evaluated only when \code{estimates} is a user-supplied object)}

\item{interactive}{A logical flag indicating whether the plotting function should
produce an interactive ShinyApp or a static ggplot2 object (the default)}

\item{column_vis}{Show color palette in a legend}
}
\description{
Plots the lower triangular visualization
based on a provided matrix
}
