% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point_estimates.R
\name{plot_point_estimates}
\alias{plot_point_estimates}
\title{Generate (interactive) plots for multi-group comparisons of point estimates}
\usage{
plot_point_estimates(
  estimates,
  type = c("analytical", "simulation", "bootstrap", "bayesian"),
  round_dec = 2,
  color_palette = "viridis",
  p_bars = FALSE,
  one_tailed_test = FALSE,
  p_val_threshold = 0.05,
  p_val_type = "p-value",
  interactive = FALSE,
  caption_exp = "Expected values",
  caption_stat = "Statistical Significance of\\nPairwise Differences",
  add_vline = NULL,
  title = NULL,
  xlim = NULL
)
}
\arguments{
\item{estimates}{A \code{triviz_estimates} object generated by the package's suite
of \code{get_est} commands or a user-supplied object of equivalent structure (for
details, please see the vignette)}

\item{type}{Specifies how uncertainty estimates are derived (evaluated only when
\code{estimates} is a user-supplied object). Input must be one of
\code{"analytical"}, \code{"simulation"}, \code{"bootstrap"}, or \code{"bayesian"}.}

\item{round_dec}{Number of decimal points for rounding of the x-axis}

\item{color_palette}{Character specifying the name of the color palette to be used.
Defaults to \code{"viridis"}.}

\item{p_bars}{Show p-values as bars, not as shadings.}

\item{one_tailed_test}{= (evaluated only when \code{estimates} is a user-supplied object)}

\item{p_val_threshold}{= (evaluated only when \code{estimates} is a user-supplied object)}

\item{p_val_type}{(evaluated only when \code{estimates} is a user-supplied object)}

\item{interactive}{A logical flag indicating whether the plotting function should
produce an interactive ShinyApp or a static ggplot2 object (the default)}

\item{caption_exp}{A custom caption for the left-hand side of the plot.}

\item{caption_stat}{A custom caption for the right-hand side of the plot.}

\item{add_vline}{Optional: A numerical value for a red vertical line in the
left-hand side of the plot.}

\item{title}{Optional: A main title for the plot.}

\item{xlim}{Optional: A range of values for the horizontal axis of the left-hand side
of the plot.}
}
\description{
Plots expected values and triangular visualizations of their pairwise
differences based on a \code{triviz} objects
}
