% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triviz-package.R
\docType{package}
\name{triviz-package}
\alias{triviz}
\alias{triviz-package}
\title{Triangular Visualizations for the Significance of Pairwise
Differences in Multigroup Comparisons}
\description{
The \code{triviz} package offers functions for
plotting group-specific estimates along with information on the
uncertainty about all of their pairwise differences. Next to
static \code{ggplot2}-based plotting functions, it also offers
tools for the in-depth exploration of these pairwise differences
via interactive \code{shiny} Apps.
}
\details{
The 'triviz' package
}
\references{
\enumerate{
  \item Cohen, Denis (2022). Making sense of statistical significance
  in multi-group comparisons using triangular visualizations.
  Unpublished manuscript.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/denis-cohen/triviz}
  \item Report bugs at \url{https://github.com/denis-cohen/triviz/issues}
}

}
\author{
\strong{Maintainer}: Denis Cohen \email{denis.cohen@mzes.uni-mannheim.de}

Authors:
\itemize{
  \item Andreas K<U+00FC>pfer \email{akuepfer@mail.uni-mannheim.de}
}

}
\keyword{internal}
